scrapers/requirements.txt
httpx
beautifulsoup4
playwright
feedparser
sentence-transformers
psycopg2-binary
python-dotenv
pdfplumber
